% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sim_mirt}
\alias{data_sim_mirt}
\title{Simulated dataset for multivariate item response theory model.}
\format{An object of class \code{list} of length 9.}
\usage{
data_sim_mirt
}
\description{
The dataset contains the simulation setting and the response data.
}
\keyword{datasets}
