% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sim_pcirt}
\alias{data_sim_pcirt}
\title{Simulated dataset for generalized partial credit model.}
\format{An object of class \code{list} of length 10.}
\usage{
data_sim_pcirt
}
\description{
The dataset contains the simulation setting and the response data.
}
\keyword{datasets}
